app/requirements.txt
psycopg2-binary
xarray
cfgrib
eccodes
numpy
